requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.0
attrs==24.2.0
beautifulsoup4==4.12.3
boto3==1.35.53
botocore==1.35.53
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
coloredlogs==15.0.1
dataclasses-json==0.6.7
dill==0.3.9
distro==1.9.0
fastembed==0.4.1
filelock==3.16.1
flatbuffers==24.3.25
free_proxy==1.1.2
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
googlesearch-python==1.2.5
greenlet==3.1.1
groq==0.11.0
grpcio==1.67.1
grpcio-status==1.67.1
grpcio-tools==1.67.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2024.2.26
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
jiter==0.7.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.6
langchain-aws==0.2.6
langchain-community==0.3.4
langchain-core==0.3.15
langchain-google-genai==2.0.4
langchain-groq==0.2.1
langchain-mistralai==0.2.1
langchain-ollama==0.2.0
langchain-openai==0.2.5
langchain-text-splitters==0.3.1
langsmith==0.1.139
loguru==0.7.2
lxml==5.3.0
magic-filter==1.0.12
marshmallow==3.23.0
minify_html==0.15.0
mistral_common==1.4.4
mmh3==4.1.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
numpy==1.26.4
ollama==0.3.3
onnx==1.17.0
onnxruntime==1.19.2
openai==1.53.0
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pillow==10.4.0
playwright==1.48.0
portalocker==2.10.1
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
py_rust_stemmers==0.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyee==12.0.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.12.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.20.1
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
scrapegraphai==1.28.0
semchunk==2.2.0
sentencepiece==0.2.0
setuptools==75.3.0
simpleeval==1.0.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.7.0
tokenizers==0.20.1
tqdm==4.66.6
transformers==4.46.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
undetected-playwright==0.3.0
uritemplate==4.1.1
urllib3==2.2.3
validators==0.34.0
yarl==1.17.1
